[exporter/datadog] Move translator and attributes into `model`
