+++ exporter/datadogexporter/internal/model/internal/utils/tags_test.go
@@ -12,20 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package translator // import "github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/translator"
+package utils // import "github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/model/internal/utils"
 
-import "context"
+import (
+	"testing"
 
-// HostnameProvider gets a hostname
-type HostnameProvider interface {
-	// Hostname gets the hostname from the machine.
-	Hostname(ctx context.Context) (string, error)
-}
+	"github.com/stretchr/testify/assert"
+)
 
-var _ HostnameProvider = (*noHostProvider)(nil)
+func TestFormatKeyValueTag(t *testing.T) {
+	tests := []struct {
+		key         string
+		value       string
+		expectedTag string
+	}{
+		{"a.test.tag", "a.test.value", "a.test.tag:a.test.value"},
+		{"a.test.tag", "", "a.test.tag:n/a"},
+	}
 
-type noHostProvider struct{}
-
-func (*noHostProvider) Hostname(context.Context) (string, error) {
-	return "", nil
+	for _, testInstance := range tests {
+		assert.Equal(t, testInstance.expectedTag, FormatKeyValueTag(testInstance.key, testInstance.value))
+	}
 }
